mass rename message_util files
