+++ include/message_util.h
