+++ src/battle/ov16_0226871C.c
@@ -44,7 +44,7 @@
 #include "unk_02005474.h"
 #include "unk_02006E3C.h"
 #include "message.h"
-#include "unk_0200B29C.h"
+#include "message_util.h"
 #include "string_template.h"
 #include "unk_0200C6E4.h"
 #include "unk_0200D9E8.h"
