+++ src/message_util.c
@@ -5,7 +5,7 @@
 #include "strbuf.h"
 
 #include "message.h"
-#include "unk_0200B29C.h"
+#include "message_util.h"
 #include "string_template.h"
 #include "strbuf.h"
