+++ src/overlay005/ov5_021E1D20.c
@@ -16,7 +16,7 @@
 
 #include "unk_02002B7C.h"
 #include "message.h"
-#include "unk_0200B29C.h"
+#include "message_util.h"
 #include "string_template.h"
 #include "unk_0200DA60.h"
 #include "heap.h"
