+++ src/overlay021/ov21_021E737C.c
@@ -31,7 +31,7 @@
 #include "unk_02009714.h"
 #include "unk_0200A328.h"
 #include "message.h"
-#include "unk_0200B29C.h"
+#include "message_util.h"
 #include "unk_0200D9E8.h"
 #include "unk_02015064.h"
 #include "heap.h"
