+++ src/overlay056/ov56_022561C0.c
@@ -37,7 +37,7 @@
 #include "unk_02009714.h"
 #include "unk_0200A328.h"
 #include "message.h"
-#include "unk_0200B29C.h"
+#include "message_util.h"
 #include "string_template.h"
 #include "unk_02014A84.h"
 #include "unk_02017728.h"
