+++ src/overlay088/ov88_0223B140.c
@@ -54,7 +54,7 @@
 #include "unk_0200A784.h"
 #include "unk_0200A9DC.h"
 #include "message.h"
-#include "unk_0200B29C.h"
+#include "message_util.h"
 #include "string_template.h"
 #include "unk_0200DA60.h"
 #include "unk_0200F174.h"
