+++ src/overlay088/ov88_0223E9C4.c
@@ -17,7 +17,7 @@
 #include "unk_02001AF4.h"
 #include "unk_02002B7C.h"
 #include "unk_02005474.h"
-#include "unk_0200B29C.h"
+#include "message_util.h"
 #include "unk_0200DA60.h"
 #include "unk_02018340.h"
 #include "unk_0201D670.h"
