+++ src/overlay094/ov94_0223DC04.c
@@ -21,7 +21,7 @@
 #include "unk_02005474.h"
 #include "unk_02006E3C.h"
 #include "message.h"
-#include "unk_0200B29C.h"
+#include "message_util.h"
 #include "string_template.h"
 #include "unk_0200DA60.h"
 #include "unk_0200F174.h"
