+++ src/overlay097/ov97_0222D30C.c
@@ -42,7 +42,7 @@
 #include "unk_02009714.h"
 #include "unk_0200A328.h"
 #include "message.h"
-#include "unk_0200B29C.h"
+#include "message_util.h"
 #include "string_template.h"
 #include "unk_0200D9E8.h"
 #include "unk_0200DA60.h"
