+++ src/overlay109/ov109_021D3D50.c
@@ -35,7 +35,7 @@
 #include "unk_0200A328.h"
 #include "unk_0200A784.h"
 #include "message.h"
-#include "unk_0200B29C.h"
+#include "message_util.h"
 #include "string_template.h"
 #include "unk_0200D9E8.h"
 #include "unk_0200DA60.h"
