+++ src/unk_02052C6C.c
@@ -22,7 +22,7 @@
 
 #include "unk_02005474.h"
 #include "message.h"
-#include "unk_0200B29C.h"
+#include "message_util.h"
 #include "string_template.h"
 #include "unk_0200DA60.h"
 #include "unk_0200F174.h"
